requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.2
dnspython==2.3.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
